fix: usePlatform으로 네이밍 변경 및 userAgent에 sopt 키워드 추가 (#1245)
